pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "surfaces-agent"
version = "0.1.0"
description = "Autonomous agentic AI engine for surface chemistry and catalytic modeling"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "pydantic>=2.0",
    "ase>=3.22",
    "pymatgen>=2023.10",
    "google-genai",
    "openai",
    "mp-api",
    "python-dotenv",
    "chgnet",
    "matplotlib",
    "numpy",
    "scipy"
]

[project.scripts]
surfaces-agent = "surfaces_agent.agent.engine:main"
surfaces-slab = "surfaces_agent.tools.slab:main"
surfaces-mp = "surfaces_agent.tools.mp:main"
surfaces-pourbaix = "surfaces_agent.tools.pourbaix:main"
surfaces-save = "surfaces_agent.tools.io:main"
surfaces-adsorb = "surfaces_agent.tools.adsorption:main"
surfaces-analyze = "surfaces_agent.tools.analysis:main"
surfaces-md = "surfaces_agent.tools.md:main"
surfaces-search = "surfaces_agent.tools.search:main"
surfaces-supercell = "surfaces_agent.tools.supercell:main"

[tool.setuptools]
packages = ["surfaces_agent", "surfaces_agent.agent", "surfaces_agent.tools"]